% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptiviti_frameworks.R
\name{receptiviti_frameworks}
\alias{receptiviti_frameworks}
\title{List Available Frameworks}
\usage{
receptiviti_frameworks(url = Sys.getenv("RECEPTIVITI_URL"),
  key = Sys.getenv("RECEPTIVITI_KEY"),
  secret = Sys.getenv("RECEPTIVITI_SECRET"))
}
\arguments{
\item{url, key, secret}{Request arguments; same as those in \code{\link{receptiviti}}.}
}
\value{
A character vector containing the names of frameworks available to your account.
}
\description{
Retrieve the list of frameworks available to your account.
}
\examples{
\dontrun{

# see which frameworks are available to your account
frameworks <- receptiviti_frameworks()
}
}
