% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptiviti.R
\name{receptiviti}
\alias{receptiviti}
\title{Receptiviti API}
\usage{
receptiviti(text, output = NULL, key = Sys.getenv("RECEPTIVITI_KEY"),
  secret = Sys.getenv("RECEPTIVITI_SECRET"),
  url = Sys.getenv("RECEPTIVITI_URL"), retry_limit = 10, cache = TRUE,
  verbose = FALSE)
}
\arguments{
\item{text}{A character vector with text to be processed.}

\item{output}{Path to a \code{.csv} file to write results to. If this already exists, it will be loaded instead of
processing any text.}

\item{key}{API Key; defaults to \code{Sys.getenv("RECEPTIVITI_KEY")}.}

\item{secret}{API Secret; defaults to \code{Sys.getenv("RECEPTIVITI_SECRET")}.}

\item{url}{API endpoint; defaults to \code{Sys.getenv("RECEPTIVITI_URL")}, which defaults to
\code{"https://api.receptiviti.com/"}.}

\item{retry_limit}{Maximum number of times each request can be retried after hitting a rate limit.}

\item{cache}{Logical; if \code{FALSE}, will not temporarily store responses, to avoid making the same requests multiple times.}

\item{verbose}{Logical; if \code{TRUE}, will show status messages.}
}
\value{
A \code{data.frame} with columns for \code{text} (the originally entered text), \code{response_id}, \code{language},
and \code{version} (metadata from the API), and scores from each included framework (e.g., \code{summary.word_count} and \code{liwc.i}).
}
\description{
The main function to access the \href{https://www.receptiviti.com}{Receptiviti} API.
}
\examples{
\dontrun{
# score a single text
single <- receptiviti("a text to score")

# score multiple texts, and write results to a file
multi <- receptiviti(c("first text to score", "second text"), "filename.csv")
}
}
